added transparent platform; modified/removed collision "fixer" script

+++ SE101 Unity Physics/Assets/Scripts/Tilt.cs
@@ -33,11 +33,11 @@ public class Tilt : MonoBehaviour
         {
             if (Input.GetAxis("Horizontal") > .2)
             {
-                transform.Rotate(0, 0, (float)0.5);
+                transform.Rotate(0, 0, 0.5f);
             }
             if (Input.GetAxis("Horizontal") < -.2)
             {
-                transform.Rotate(0, 0, (float)-0.5);
+                transform.Rotate(0, 0, -0.5f);
             }
             if (Input.GetAxis("Vertical") > .2)
             {
